First Pass at adding NetImgui
Current implementation makes game unusable while Cog is open in NetImgui, and NetImgui when Cog is not opened in game viewport. So not ideal, but it does work!
Would also like to add the option to disable NetImgui calls at compile time
Added NetImgui library and Server app along with Licenses